% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est-SAEM_USER.R
\name{predict.sclmSPDE}
\alias{predict.sclmSPDE}
\title{Prediction in spatial models with censored/missing responses}
\usage{
\method{predict}{sclmSPDE}(object, x.new, coord.new, ...)
}
\arguments{
\item{object}{object of class \code{'sclmSPDE'} given as output of \code{\link{SPDEsclm}} function.}

\item{x.new}{matrix of covariates for which prediction is performed.}

\item{coord.new}{matrix of coordinates for which prediction is performed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function returns a list with:
\item{coord}{matrix of coordinates.}
\item{predValues}{predicted values.}
\item{sdPred}{predicted standard deviations.}
}
\description{
It performs spatial prediction in a set of new \code{S} spatial locations.
}
\details{
This function predicts using the mean squared error (MSE) criterion, which
takes the conditional expectation E(Y|X) as the best linear predictor.
}
\examples{
\donttest{library("RcppCensSpatial")
library("SPDEcensSpatial")

set.seed(1000)
n = 120
coords = round(matrix(runif(2*n,0,15),n,2), 5)
x = cbind(rbinom(n,1,0.50), rnorm(n), rnorm(n))
data = rCensSp(c(1,4,-1), 2, 3, 0.50, x, coords, "left", 0.10, 20, "matern", 1)

## Estimation
data1 = data$Data

# Estimation: SAEM algorithm
fit = SPDEsclm(y=data1$y, x=data1$x, ci=data1$ci, lcl=data1$lcl,
               ucl=data1$ucl, coords=data1$coords, phi0=2.5,
               lower = c(0.01, 0.01), upper = c(30, 0.99), iedge = 0.3)
fit$tab

# Prediction
data2 = data$TestData
pred = predict(fit, data2$x, data2$coords)

# Cross-validation
mean((data2$y - pred$predValues)^2)}
}
\seealso{
\code{\link{SPDEsclm}}
}
\author{
Pablo Zúñiga, Katherine L. Valeriano, and Lourival Lima
}
